+++ tests/compressor_deviation_test.py
@@ -1,12 +1,10 @@
 import unittest
-from turbodesigner.blade import BladeDeviation, AirfoilType
+from turbodesigner.blade.row import BladeDeviation, AirfoilType
 import numpy as np
 
-class CascadeJohnsonDesignTest(unittest.TestCase):
+class AungierDeviationTest(unittest.TestCase):
 
     def test_aungier_deviation(self):
-
-
         deviation = BladeDeviation(
             beta1=np.radians(70),               # rad
             beta2=np.radians(20),               # rad
@@ -17,8 +15,8 @@ class CascadeJohnsonDesignTest(unittest.TestCase):
 
         metal_angles = deviation.get_metal_angles(100)
 
-        self.assertAlmostEqual(np.degrees(metal_angles.kappa1), 73.9657, 4)
-        self.assertAlmostEqual(np.degrees(metal_angles.kappa2), -0.4597, 4)
+        np.testing.assert_almost_equal(np.degrees(metal_angles.kappa1), 73.9657, 4)
+        np.testing.assert_almost_equal(np.degrees(metal_angles.kappa2), -0.4597, 4)
 
 
 if __name__ == '__main__':
