+++ turbodesigner/airfoils/__init__.py
@@ -1,0 +1,2 @@
+from turbodesigner.airfoils.common import AirfoilType
+from turbodesigner.airfoils.DCA import DCAAirfoil
