+++ turbodesigner/units.py
@@ -1,0 +1,1 @@
+MM = 1/1000
